+++ src/main/java/com/github/zhuyizhuo/generator/mybatis/database/mapper/MysqlDataBaseMapper.java
@@ -7,6 +7,7 @@ import com.github.zhuyizhuo.generator.mybatis.database.pojo.DbTableInfo;
 import java.util.List;
 
 /**
+ *  mysql 数据库 mapper
  * @author yizhuo
  * @version 1.0
  * time: 2018/7/27 20:55
