+++ src/main/java/com/github/zhuyizhuo/generator/mybatis/database/mapper/OracleDataBaseMapper.java
@@ -8,6 +8,7 @@ import org.apache.ibatis.annotations.Param;
 import java.util.List;
 
 /**
+ *  oracle 数据库 mapper
  * @author yizhuo
  * @version 1.0
  * time: 2018/7/27 20:55
