+++ src/main/java/com/github/zhuyizhuo/generator/mybatis/database/service/impl/MysqlDbServiceImpl.java
@@ -16,7 +16,7 @@ import java.util.List;
 
 /**
  * class: MysqlDbServiceImpl <br>
- * description: mysql数据库查询表结构实现 <br>
+ * description: mysql 数据库查询表结构实现 <br>
  * time: 2018/7/30 13:10
  *
  * @author yizhuo <br>
