+++ src/main/java/com/github/zhuyizhuo/generator/mybatis/dto/JavaColumnInfo.java
@@ -15,20 +15,10 @@ public class JavaColumnInfo extends ColumnInfo {
 	private String javaColumnName;
 	/** java字段类型 */
 	private String javaDataType;
-	/** java字段类型全称 TODO java.lang 包下的未赋值 */
+	/** java字段类型全称 */
 	private String javaDataTypeFullPath;
 
 	public JavaColumnInfo() {
-	}
-
-	public JavaColumnInfo(JavaColumnInfo javaColumnInfo) {
-		super(javaColumnInfo);
-		if (javaColumnInfo == null){
-			return;
-		}
-		this.javaColumnName = javaColumnInfo.getJavaColumnName();
-		this.javaDataType = javaColumnInfo.getJavaDataType();
-		this.javaDataTypeFullPath = javaColumnInfo.getJavaDataTypeFullPath();
 	}
 
 	public String getJavaColumnName() {
