+++ src/main/java/com/github/zhuyizhuo/generator/mybatis/enums/FileTypeEnums.java
@@ -1,5 +1,10 @@
 package com.github.zhuyizhuo.generator.mybatis.enums;
 
+/**
+ * 文件类型枚举
+ * 2019年5月26日22:14:22
+ * @version 1.4.0
+ */
 public enum FileTypeEnums {
     JAVA,
     XML,
