--- src/main/java/com/github/zhuyizhuo/generator/mybatis/enums/XmlNameFormatEnums.java
@@ -1,17 +1,0 @@
-package com.github.zhuyizhuo.generator.mybatis.enums;
-
-/**
- * class: XmlNameFormatEnums <br>
- * description: TODO <br>
- * time: 2019/5/22
- *
- * @author yizhuo <br>
- * @since #since#
- */
-public enum XmlNameFormatEnums {
-    /** 表名称小写 */
-    LOWER_CASE,
-    /** 表名称转驼峰 */
-    CAMEL,
-    ;
-}
