+++ src/main/java/com/github/zhuyizhuo/generator/mybatis/generator/service/GenerateService.java
@@ -1,14 +1,14 @@
-package com.github.zhuyizhuo.generator.mybatis.service;
+package com.github.zhuyizhuo.generator.mybatis.generator.service;
 
 import com.github.zhuyizhuo.generator.mybatis.vo.GenerateMetaData;
 
 /**
  * class: GenerateService <br>
- * description: TODO <br>
+ * description: 生成器 service <br>
  * time: 2019/5/23
  *
  * @author yizhuo <br>
- * @since #since#
+ * @since 1.4.0
  */
 public interface GenerateService {
