+++ src/main/resources/freemarker/template/common/xml/deleteByWhere.ftl
@@ -1,6 +1,5 @@
     <!-- ${methodDescription.DELETE_BY_WHERE.comment} -->
 	<delete id="${methodDescription.DELETE_BY_WHERE.methodName}" parameterType="${mybatisXmlDefinition.parameterType}">
-        DELETE FROM
-        ${tableInfo.tableSchema}.${tableInfo.tableName}
+        DELETE FROM ${tableInfo.tableSchema}.${tableInfo.tableName}
         <include refid="Where_Clause" />
 	</delete>
