+++ src/main/resources/freemarker/template/common/xml/queryByPrimaryKey.ftl
@@ -6,8 +6,7 @@
 </#if>
         SELECT
         <include refid="Base_Column_List" />
-  	     FROM
-        ${tableInfo.tableSchema}.${tableInfo.tableName}
+  	     FROM ${tableInfo.tableSchema}.${tableInfo.tableName}
         WHERE
         <#list tableInfo.primaryKeyColumns as colm>
             <#if colm_index != 0>AND </#if>${colm.columnName} = ${'#{'}${colm.javaColumnName}}
