+++ src/main/resources/freemarker/template/common/xml/queryByWhere.ftl
@@ -2,7 +2,6 @@
 	<select id="${methodDescription.QUERY_BY_WHERE.methodName}" resultMap="${mybatisXmlDefinition.resultMap.id}" parameterType="${mybatisXmlDefinition.parameterType}">
         SELECT
         <include refid="Base_Column_List" />
-  	     FROM
-        ${tableInfo.tableSchema}.${tableInfo.tableName}
+  	     FROM ${tableInfo.tableSchema}.${tableInfo.tableName}
         <include refid="Where_Clause" />
     </select>
