+++ src/main/resources/freemarker/template/common/xml/updateByPrimaryKey.ftl
@@ -1,7 +1,6 @@
     <!-- ${methodDescription.UPDATE_BY_PRIMARY_KEY.comment} -->
     <update id="${methodDescription.UPDATE_BY_PRIMARY_KEY.methodName}" parameterType="${mybatisXmlDefinition.parameterType}">
-		UPDATE
-		${tableInfo.tableSchema}.${tableInfo.tableName}
+		UPDATE ${tableInfo.tableSchema}.${tableInfo.tableName}
 		<set>
     <#list tableInfo.columnLists as colm>
         <#if colm??>
